+++ Examples/OpcPlcManager/PlcManager.xml
