--- Examples/OpcUaMethodExample/src/OpcPlcManagerLibrary.acf.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<AcfConfigurationDocument
-  xmlns="http://www.phoenixcontact.com/schema/acfconfig"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://www.phoenixcontact.com/schema/acfconfig.xsd"
-  schemaVersion="1.0" >
-
-  <Processes>
-    <Process name="OpcPlcManagerProcess" settingsPath="$ARP_ACF_SETTINGS_FILE$" />
-  </Processes>
-
-  <Libraries>
-    <Library name="OpcPlcManager.OpcPlcManagerLibrary" binaryPath="../OpcPlcManager/lib/libOpcPlcManager.so" />
-  </Libraries>
-
-  <Components>
-    <Component name="OpcPlcManagerComponent1" type="OpcPlcManager::OpcPlcManagerComponent" library="OpcPlcManager.OpcPlcManagerLibrary" />
-  </Components>
-
-</AcfConfigurationDocument>
